Refactor MachineDetailsFragment to use Muscle enum
This change introduces a new set of selectedMuscles, which takes over
from the musclesArray list and _selections array that was used in the
previous version. selectedMuscles is now the single source of truth for
which muscles the user has selected: checking if a muscle is selected is
now as simple as checking if it is in the selectedMuscles set.

To limit the changes to the rest of the code, added convenience
functions that translate between sets of muscles and the bodypart string
that is used by the database.